requirements.txt
langchain==0.3.26
flask==3.1.1
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0
langchain-pinecone==0.2.13
langchain-anthropic==0.3.7
langchain-community==0.3.26
pinecone
anthropic
transformers>=4.30.0
torch>=2.0.0
pdf2image>=1.16.0
Pillow>=10.0.0
rank-bm25==0.2.2
jieba==0.42.1